+++ test.js
@@ -1,0 +1,22 @@
+import fs from 'fs'
+import nock from 'nock'
+
+const mockDataFilename = './mock-data.json'
+
+before(() => {
+  if (process.env.RECORD_MOCK_DATA) {
+    nock.recorder.rec({
+      output_objects: true,
+      dont_print: true,
+    })
+  } else {
+    nock.load(mockDataFilename)
+  }
+})
+
+after(() => {
+  if (process.env.RECORD_MOCK_DATA) {
+    const nockCallObjects = nock.recorder.play()
+    fs.writeFileSync(mockDataFilename, JSON.stringify(nockCallObjects, null, 2))
+  }
+})
