--- test/index.js
@@ -1,3 +1,0 @@
-require('babel-core/register')
-
-require('./integration/endpoints')
