--- test/run_integration_test_single.js
@@ -1,3 +1,0 @@
-require('babel-core/register')
-
-require('./integration/run_integration_test_single')
